// src/modules/news/_apis/queries/protos/news_posts_query.ts
export interface GetAllNewsPostsRequest {
    /**
     * @generated from protobuf field: optional SkipTakeOption SkipTakeOption = 1
     */
    skipTakeOption?: SkipTakeOption;
    /**
     * @generated from protobuf field: optional PagingOption PagingOption = 2
     */
    pagingOption?: PagingOption;
    /**
     * @generated from protobuf field: repeated OrderByOption OrderBy = 3
     */
    orderBy: OrderByOption[];
    /**
     * @generated from protobuf field: optional google.protobuf.Int32Value CategoryId = 4
     */
    categoryId?: Int32Value;
}
/**
 * @generated from protobuf message NewsPosts.GetDetailNewsPostsRequest
 */
export interface GetDetailNewsPostsRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.Int32Value Id = 1
     */
    id?: Int32Value;
}
// ...
class GetAllNewsPostsRequest$Type extends MessageType<GetAllNewsPostsRequest> {
    constructor() {
        super("NewsPosts.GetAllNewsPostsRequest", [
            { no: 1, name: "SkipTakeOption", kind: "message", jsonName: "SkipTakeOption", T: () => SkipTakeOption },
            { no: 2, name: "PagingOption", kind: "message", jsonName: "PagingOption", T: () => PagingOption },
            { no: 3, name: "OrderBy", kind: "message", jsonName: "OrderBy", repeat: 2 /*RepeatType.UNPACKED*/, T: () => OrderByOption },
            { no: 4, name: "CategoryId", kind: "message", jsonName: "CategoryId", T: () => Int32Value }
        ]);
    }
    create(value?: PartialMessage<GetAllNewsPostsRequest>): GetAllNewsPostsRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.orderBy = [];
        if (value !== undefined)
            reflectionMergePartial<GetAllNewsPostsRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetAllNewsPostsRequest): GetAllNewsPostsRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* optional SkipTakeOption SkipTakeOption */ 1:
                    message.skipTakeOption = SkipTakeOption.internalBinaryRead(reader, reader.uint32(), options, message.skipTakeOption);
                    break;
                case /* optional PagingOption PagingOption */ 2:
                    message.pagingOption = PagingOption.internalBinaryRead(reader, reader.uint32(), options, message.pagingOption);
                    break;
                case /* repeated OrderByOption OrderBy */ 3:
                    message.orderBy.push(OrderByOption.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* optional google.protobuf.Int32Value CategoryId */ 4:
                    message.categoryId = Int32Value.internalBinaryRead(reader, reader.uint32(), options, message.categoryId);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetAllNewsPostsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* optional SkipTakeOption SkipTakeOption = 1; */
        if (message.skipTakeOption)
            SkipTakeOption.internalBinaryWrite(message.skipTakeOption, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* optional PagingOption PagingOption = 2; */
        if (message.pagingOption)
            PagingOption.internalBinaryWrite(message.pagingOption, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* repeated OrderByOption OrderBy = 3; */
        for (let i = 0; i < message.orderBy.length; i++)
            OrderByOption.internalBinaryWrite(message.orderBy[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* optional google.protobuf.Int32Value CategoryId = 4; */
        if (message.categoryId)
            Int32Value.internalBinaryWrite(message.categoryId, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message NewsPosts.GetAllNewsPostsRequest
 */
export const GetAllNewsPostsRequest = new GetAllNewsPostsRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetDetailNewsPostsRequest$Type extends MessageType<GetDetailNewsPostsRequest> {
    constructor() {
        super("NewsPosts.GetDetailNewsPostsRequest", [
            { no: 1, name: "Id", kind: "message", jsonName: "Id", T: () => Int32Value }
        ]);
    }
    create(value?: PartialMessage<GetDetailNewsPostsRequest>): GetDetailNewsPostsRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<GetDetailNewsPostsRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetDetailNewsPostsRequest): GetDetailNewsPostsRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* optional google.protobuf.Int32Value Id */ 1:
                    message.id = Int32Value.internalBinaryRead(reader, reader.uint32(), options, message.id);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetDetailNewsPostsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* optional google.protobuf.Int32Value Id = 1; */
        if (message.id)
            Int32Value.internalBinaryWrite(message.id, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message NewsPosts.GetDetailNewsPostsRequest
 */
export const GetDetailNewsPostsRequest = new GetDetailNewsPostsRequest$Type();